requirements.txt
paramiko
testinfra
pytest-xdist
pyaml
httpie
fabric
retrying
pybookshelf
docker-compose
timeout_decorator